+++ src/main/java/br/com/medcare/model/PacienteRequest.java
@@ -1,5 +1,7 @@
 package br.com.medcare.model;
 
+import java.math.BigInteger;
+
 import lombok.Data;
 
 @Data
@@ -8,7 +10,7 @@ public class PacienteRequest {
 	    private Integer idade;
 	    private Integer telefone;
 	    private String endereco;
-	    private Integer cpf;
+	    private BigInteger cpf;
 	    private String nome;
 	    private String email;	    
 	    private String password;
